Move JSONAPI into its own file
